Convert models/IrcAction to Typescript
